Integração front e back

+++ educacaoParaTodes/src/main/java/com/educadamente/educacaoParaTodes/repository/UsuarioRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import com.educadamente.educacaoParaTodes.model.Usuario;
 
 @Repository
-public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
+public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
 	
 	 Optional<Usuario> findByEmail(String email);
 }
